requirements.txt
Django==3.0
djangorestframework==3.10.3
drf-yasg==1.17.0
factory-boy==2.12.0
Faker==3.0.0
